DAO para entidades Factura, Pago
LineaFactura no necesita un DAO, pues se mapea como atributo de Factura.

+++ src/main/java/es/uvigo/esei/dagss/facturaaas/daos/FacturaDAO.java
@@ -5,9 +5,16 @@ import es.uvigo.esei.dagss.facturaaas.entidades.Factura;
 import java.util.List;
 
 /**
+ * Especifica el contrato que debe de cumplir un DAO de Facturas.
  *
  * @author Alejandro González García
  */
-public interface FacturaDAO extends GenericoDAO<Factura, Integer> {
+public interface FacturaDAO extends GenericoDAO<Factura, Long> {
+    /**
+     * Recupera las facturas asociadas a un determinado cliente.
+     *
+     * @param cliente El cliente cuyas facturas recuperar.
+     * @return Una lista con las facturas a nombre del cliente.
+     */
     public List<Factura> buscarPorCliente(Cliente cliente);
 }
